webapi/tct-alarm-tizen-tests/alarm/AlarmRelative_constructor_delay.html

<!DOCTYPE html>

<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

*Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
*Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
*Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jenny Cao <jenny.q.cao@intel.com>
        Mariusz Polasinski <m.polasinski@samsung.com>

-->

<html>
<head>
<title>AlarmRelative_constructor_delay</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: AlarmRelative_constructor_delay
//==== LABEL Test whether the constructor with delay parameter are supported
//==== SPEC Tizen Web API:Application:Alarm:AlarmRelative:constructor C
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/alarm.html
//==== PRIORITY P1
//==== TEST_CRITERIA CONSTRA CONSTRM

test(function () {
    var alarm, delay = 3600;
    alarm = new tizen.AlarmRelative(delay);
    assert_true(alarm instanceof tizen.AlarmRelative, "object was not created properly");
    assert_equals(alarm.delay, delay, "delay passed as constructor parameter is different than delay in the object");
}, document.title);

</script>
</body>
</html>
